user menu landing polished
